attrocious quality of training deck - allow large file uploads too in every feature
multiple file upload to be supported in knowledge base update

also the created training deck dashboard feature is still missing - upload of files to create training deck is also still missing

ensure all features are there as widgets on the homepage (even the new ones that have been created and will be created in the future too)

data analysis is still completely incorrect - I want a data analysis and not a data analysis scope

also allow option to view the knowledge base entries in the knowledge base table and while creating deck too

also add option to download in the training deck dashboard and in the data analysis dashboard

also in the create training deck feature, you need to be able to create and share downloadable version of the format selected. Brochure as a pdf but like a brochure with proper graphic images and designs to be refered to from the uploaded file as input as well. Dont just share the format - share the final output.

in data analysis all excel file formats to be supported

also ensure that all output files which are currentlyt coming as .txt in this application are word docs (.doc or.docs) and not .txt